LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-buttons-styles-expected.txt

Testing the MacOSInlineMediaControls computed styles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PlayPauseButton
PASS getComputedStyle(button.element).webkitMaskPositionY became "12px"
PASS rgba(getComputedStyle(button.element).backgroundColor).r is 255
PASS rgba(getComputedStyle(button.element).backgroundColor).g is 255
PASS rgba(getComputedStyle(button.element).backgroundColor).b is 255
PASS rgba(getComputedStyle(button.element).backgroundColor).a is within 0.001 of 0.45
PASS getComputedStyle(button.element).height is "50px"

SkipBackButton
PASS getComputedStyle(button.element).webkitMaskPositionY became "10px"
PASS rgba(getComputedStyle(button.element).backgroundColor).r is 255
PASS rgba(getComputedStyle(button.element).backgroundColor).g is 255
PASS rgba(getComputedStyle(button.element).backgroundColor).b is 255
PASS rgba(getComputedStyle(button.element).backgroundColor).a is within 0.001 of 0.45
PASS getComputedStyle(button.element).height is "50px"

MuteButton
PASS getComputedStyle(button.element).webkitMaskPositionY became "10px"
PASS rgba(getComputedStyle(button.element).backgroundColor).r is 255
PASS rgba(getComputedStyle(button.element).backgroundColor).g is 255
PASS rgba(getComputedStyle(button.element).backgroundColor).b is 255
PASS rgba(getComputedStyle(button.element).backgroundColor).a is within 0.001 of 0.45
PASS getComputedStyle(button.element).height is "50px"

AirplayButton
PASS getComputedStyle(button.element).webkitMaskPositionY became "13px"
PASS rgba(getComputedStyle(button.element).backgroundColor).r is 255
PASS rgba(getComputedStyle(button.element).backgroundColor).g is 255
PASS rgba(getComputedStyle(button.element).backgroundColor).b is 255
PASS rgba(getComputedStyle(button.element).backgroundColor).a is within 0.001 of 0.45
PASS getComputedStyle(button.element).height is "50px"

PiPButton
PASS getComputedStyle(button.element).webkitMaskPositionY became "13px"
PASS rgba(getComputedStyle(button.element).backgroundColor).r is 255
PASS rgba(getComputedStyle(button.element).backgroundColor).g is 255
PASS rgba(getComputedStyle(button.element).backgroundColor).b is 255
PASS rgba(getComputedStyle(button.element).backgroundColor).a is within 0.001 of 0.45
PASS getComputedStyle(button.element).height is "50px"

TracksButton
PASS getComputedStyle(button.element).webkitMaskPositionY became "15px"
PASS rgba(getComputedStyle(button.element).backgroundColor).r is 255
PASS rgba(getComputedStyle(button.element).backgroundColor).g is 255
PASS rgba(getComputedStyle(button.element).backgroundColor).b is 255
PASS rgba(getComputedStyle(button.element).backgroundColor).a is within 0.001 of 0.45
PASS getComputedStyle(button.element).height is "50px"

FullscreenButton
PASS getComputedStyle(button.element).webkitMaskPositionY became "13px"
PASS rgba(getComputedStyle(button.element).backgroundColor).r is 255
PASS rgba(getComputedStyle(button.element).backgroundColor).g is 255
PASS rgba(getComputedStyle(button.element).backgroundColor).b is 255
PASS rgba(getComputedStyle(button.element).backgroundColor).a is within 0.001 of 0.45
PASS getComputedStyle(button.element).height is "50px"
PASS successfullyParsed is true

TEST COMPLETE
